# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Admission API",
        "description": "API for managing medical patient admissions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/scoring-charts": {
            "get": {
                "tags": [
                    "ScoringChart"
                ],
                "summary": "Get all scoring charts",
                "operationId": "9b30d802be26b036e0b061f320efedca",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScoringChart"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ScoringChart"
                ],
                "summary": "Create a new scoring chart",
                "operationId": "1f0a7926ef01419c8cd011c50c06fceb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScoringChart"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Scoring Chart created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScoringChart"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/scoring-charts/{id}": {
            "get": {
                "tags": [
                    "ScoringChart"
                ],
                "summary": "Get scoring chart by ID",
                "operationId": "e9f1c987329747cdab02d19f3dbb509c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Scoring chart ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScoringChart"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chart not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ScoringChart"
                ],
                "summary": "Update APGAR scoring chart",
                "operationId": "39878bb1798e208912205f513cac9e46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Scoring chart ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScoringChart"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScoringChart"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ScoringChart"
                ],
                "summary": "Delete scoring chart",
                "operationId": "2f78d9ad4754197ec8d424cc19f4b175",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ScoringChart": {
                "required": [
                    "patient_admission_id",
                    "dateScored",
                    "heartRate",
                    "respiratory",
                    "muscleTone",
                    "reflexes",
                    "color",
                    "heartRate5",
                    "respiratory5",
                    "muscleTone5",
                    "reflexes5",
                    "color5",
                    "heartRate10",
                    "respiratory10",
                    "muscleTone10",
                    "reflexes10",
                    "color10",
                    "otherHeartRate",
                    "otherRespiratory",
                    "otherMuscleTone",
                    "otherReflexes",
                    "otherColor"
                ],
                "properties": {
                    "patient_admission_id": {
                        "type": "integer"
                    },
                    "dateScored": {
                        "type": "string",
                        "format": "date"
                    },
                    "heartRate": {
                        "type": "integer"
                    },
                    "respiratory": {
                        "type": "integer"
                    },
                    "muscleTone": {
                        "type": "integer"
                    },
                    "reflexes": {
                        "type": "integer"
                    },
                    "color": {
                        "type": "integer"
                    },
                    "heartRate5": {
                        "type": "integer"
                    },
                    "respiratory5": {
                        "type": "integer"
                    },
                    "muscleTone5": {
                        "type": "integer"
                    },
                    "reflexes5": {
                        "type": "integer"
                    },
                    "color5": {
                        "type": "integer"
                    },
                    "heartRate10": {
                        "type": "integer"
                    },
                    "respiratory10": {
                        "type": "integer"
                    },
                    "muscleTone10": {
                        "type": "integer"
                    },
                    "reflexes10": {
                        "type": "integer"
                    },
                    "color10": {
                        "type": "integer"
                    },
                    "otherHeartRate": {
                        "type": "integer"
                    },
                    "otherRespiratory": {
                        "type": "integer"
                    },
                    "otherMuscleTone": {
                        "type": "integer"
                    },
                    "otherReflexes": {
                        "type": "integer"
                    },
                    "otherColor": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}